final/requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langgraph
pandas
sqlalchemy
python-dotenv
tabulate
